# [speed] Don't init moduleFormatter unless we need it