@amasad 2015-05-07
Ran it on a real project with real transforms and it wasn't as impactful as this suggests, might be because a lot of what `moduleFormatter.init` does is cached?