@sebmck 2015-05-07
I think the reason I did this in the first place was that sometimes transformers inject modules/imports into the code. Pretty sure this would break the `runtime` and `bluebirdCoroutines` transformers if you didn't already have any imports/exports since `transformers["es6.modules"].canTransform()` would return `false` if there were no imports or exports in the tree.