@amasad 2015-05-07
Should we then change `transformers["es6.modules"].canTransform()` to be smarter?