@amasad 2015-05-07
Also, why is there an "internal" strict and regular strict transform. And the internal one runs and visits just to get to [this line](https://github.com/babel/babel/blob/master/src/babel/transformation/transformers/internal/strict.js#L4) which will basically tell it not to do anything. This takes ~250ms for jquery