@sebmck 2015-05-07
They're basically the same transformer but split up. `strict` removes any `"use strict";` directive and builds one if necessary then saves it, `_strict` then reads it and adds it once all the other transformers have ran so it's always on the top. I recall adding some directive hoisting logic somewhere (may be daydreaming, can't seem to find it now) so it may be irrelevant. Regardless, `this.stop()` can be used on it since it doesn't visit any other nodes, might be a good idea to special case single-node visitors like this in the actual traversal logic so it doesn't have to be done for every single visitor.