@amasad 2015-05-07
Cool. `this.stop()` in `_strict` seems like a couple of seconds in a 70s compile. Maybe there are other places this applies.
Let me know if there is anything we can do for this PR.