@sebmck 2015-05-07
Well this would kill dynamic imports so I can't really merge. I'm going to experiment with merging all the internal (all built-in ones) transformers into a single traversal. Not sure how much I'd have to change or how easy it'd be to maintain but it would eliminate a lot of the performance concerns that I've been trying to get around with this prepass stuff. Plugins would continue to have a traversal per to avoid conflicts since they can't exactly be aware of the transformers "around" them.

Could possibly introduce the concept of traversal "categories". ie. the built-in transformers would expose that they belong to the "category" `builtin` and then internally they'd be merged together into a single pass and plugins could either expose their own "category" (eg. `facebook`) and those would be merged together too.