@sebmck 2015-05-07
I merged all the builtin transformer visitors and ran the test suite:

```
  263 passing
  4 pending
  248 failing
```

I've been running Babel over `tsserver.js` which is a 1.8MB blob of JavaScript as a sort of stress test since it contains a lot of nodes and really puts the traversal algo to the test.

**Original**

```
~/Desktop$ DEBUG=babel time babel tsserver.js --out-file tsserver.foo.js
  babel [BABEL] tsserver.js: Parse start +0ms
  babel [BABEL] tsserver.js: Parse stop +741ms
  babel [BABEL] tsserver.js: Start set AST +1ms
  babel [BABEL] tsserver.js: Start scope building +0ms
  babel [BABEL] tsserver.js: End scope building +3s
  babel [BABEL] tsserver.js: End set AST +868ms
  babel [BABEL] tsserver.js: Start prepass +0ms
  babel [BABEL] tsserver.js: End prepass +2s
  babel [BABEL] tsserver.js: Start module formatter init +0ms
  babel [BABEL] tsserver.js: End module formatter init +1s
  babel [BABEL] tsserver.js: Start transformer strict +0ms
  babel [BABEL] tsserver.js: Finish transformer strict +3ms
  babel [BABEL] tsserver.js: Start transformer _validation +0ms
  babel [BABEL] tsserver.js: Finish transformer _validation +679ms
  babel [BABEL] tsserver.js: Start transformer validation.react +0ms
  babel [BABEL] tsserver.js: Finish transformer validation.react +737ms
  babel [BABEL] tsserver.js: Start transformer spec.functionName +0ms
  babel [BABEL] tsserver.js: Finish transformer spec.functionName +883ms
  babel [BABEL] tsserver.js: Start transformer spec.blockScopedFunctions +0ms
  babel [BABEL] tsserver.js: Finish transformer spec.blockScopedFunctions +781ms
  babel [BABEL] tsserver.js: Start transformer es6.tailCall +0ms
  babel [BABEL] tsserver.js: Finish transformer es6.tailCall +1s
  babel [BABEL] tsserver.js: Start transformer _blockHoist +0ms
  babel [BABEL] tsserver.js: Finish transformer _blockHoist +706ms
  babel [BABEL] tsserver.js: Start transformer _shadowFunctions +0ms
  babel [BABEL] tsserver.js: Finish transformer _shadowFunctions +1s
  babel [BABEL] tsserver.js: Start transformer _strict +0ms
  babel [BABEL] tsserver.js: Finish transformer _strict +1ms
  babel [BABEL] tsserver.js: Start transformer _moduleFormatter +0ms
  babel [BABEL] tsserver.js: Finish transformer _moduleFormatter +0ms
  babel [BABEL] tsserver.js: Start transformer es3.propertyLiterals +0ms
  babel [BABEL] tsserver.js: Finish transformer es3.propertyLiterals +691ms
  babel [BABEL] tsserver.js: Start transformer es3.memberExpressionLiterals +0ms
  babel [BABEL] tsserver.js: Finish transformer es3.memberExpressionLiterals +740ms
  babel [BABEL] tsserver.js: Generation start +1ms
[BABEL] Note: The code generator has deoptimised the styling of "tsserver.js" as it exceeds the max of "100KB".
  babel [BABEL] tsserver.js: Generation end +1s
       18.10 real        18.31 user         0.27 sys
```

**Combined**

```
~/Desktop$ DEBUG=babel time babel tsserver.js --out-file tsserver.foo.js
  babel [BABEL] tsserver.js: Parse start +0ms
  babel [BABEL] tsserver.js: Parse stop +905ms
  babel [BABEL] tsserver.js: Start set AST +1ms
  babel [BABEL] tsserver.js: Start scope building +1ms
  babel [BABEL] tsserver.js: End scope building +3s
  babel [BABEL] tsserver.js: End set AST +917ms
  babel [BABEL] tsserver.js: Start module formatter init +0ms
  babel [BABEL] tsserver.js: End module formatter init +1s
  babel [BABEL] tsserver.js: Start transformer builtin +1ms
  babel [BABEL] tsserver.js: Finish transformer builtin +5s
  babel [BABEL] tsserver.js: Generation start +0ms
[BABEL] Note: The code generator has deoptimised the styling of "tsserver.js" as it exceeds the max of "100KB".
  babel [BABEL] tsserver.js: Generation end +2s
       13.70 real        14.15 user         0.26 sys
```

18s to 13s Not as dramatic of an improvement as I'd hoped.

Generation can be **heavily** improved by switching the generation to `escodegen`, it goes from `~2s` to `200ms`:

```
  babel [BABEL] tsserver.js: Parse start +0ms
  babel [BABEL] tsserver.js: Parse stop +887ms
  babel [BABEL] tsserver.js: Start set AST +2ms
  babel [BABEL] tsserver.js: Start scope building +0ms
  babel [BABEL] tsserver.js: End scope building +3s
  babel [BABEL] tsserver.js: End set AST +889ms
  babel [BABEL] tsserver.js: Start module formatter init +0ms
  babel [BABEL] tsserver.js: End module formatter init +1s
  babel [BABEL] tsserver.js: Start transformer builtin +0ms
  babel [BABEL] tsserver.js: Finish transformer builtin +5s
  babel [BABEL] tsserver.js: Generation start +0ms
  babel [BABEL] tsserver.js: Generation end +217ms
```

Extremely dramatic improvement and shows how much the current code generator could be optimised.

The overhead of the initial scope tracking/buidling is also pretty shit, unsure why that's taking so long. Probably something stupid like a lot of excessive traversals.