@amasad 2015-05-07
Generation is not a bottle neck in real world applications. I assume `code: false` turns off generation? Doing that has very little impact. I should say we also use `compact: true, retainLines: true`